+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Optimized Stop Word Filter
     - Allow list of empty regex patterns in Regex Filter
     - Handle missing class definitions in Native and Igbinary
+    - Fixed infinite loop in Ball Tree & KD Tree grow method
 
 - 0.2.0
     - Add Recursive Feature Eliminator feature selector